BREAK: rename ValueSchema to IdSchema
